diploma/pet_diary/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-crispy-forms==2.3
Pillow==9.5.0
sqlparse==0.5.3
typing_extensions==4.12.2
python-dotenv~=1.0.1
python-telegram-bot~=20.3